Renamed profile route and page to writer/book route and page.
